Break lint functions out into their own files
