+++ docs/api/_src/lint_tests/files_exist.rst
@@ -1,0 +1,4 @@
+files_exist
+============
+
+.. automethod:: nf_core.lint.PipelineLint.files_exist
