+++ nf_core/lint/schema_params.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+
+import nf_core.schema
+
+
+def schema_params(self):
+    """ Check that the schema describes all flat params in the pipeline """
+    passed = []
+    warned = []
+    failed = []
+
+    # First, get the top-level config options for the pipeline
+    # Schema object already created in the `schema_lint` test
+    self.schema_obj.get_schema_path(self.path)
+    self.schema_obj.get_wf_params()
+    self.schema_obj.no_prompts = True
+
+    # Remove any schema params not found in the config
+    removed_params = self.schema_obj.remove_schema_notfound_configs()
+
+    # Add schema params found in the config but not the schema
+    added_params = self.schema_obj.add_schema_found_configs()
+
+    if len(removed_params) > 0:
+        for param in removed_params:
+            warned.append("Schema param `{}` not found from nextflow config".format(param))
+
+    if len(added_params) > 0:
+        for param in added_params:
+            failed.append("Param `{}` from `nextflow config` not found in nextflow_schema.json".format(param))
+
+    if len(removed_params) == 0 and len(added_params) == 0:
+        passed.append("Schema matched params returned from nextflow config")
+
+    return {"passed": passed, "warned": warned, "failed": failed}
